requirements.txt
Flask
python-dotenv
secrets
mysql-connector-python
pandas
smtplib
email
schedule
datetime
